+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Upgrade dependencies
 - Generate gleam files synchronously
 - Do not add comments to generated file, for more predictable LSP usage
+- Add `inject` and `provide`
 
 ## v0.1.0